GIS12: Implement PM 2012 lot 1 / Inverse geocoding / Facades
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/technical/plugin-gis@40275 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/gis/web/GisAddressJspBean.java
@@ -74,7 +74,7 @@ public class GisAddressJspBean extends PluginAdminPageJspBean
 		if ( ( x != null && x.length() != 0 )  && 
 			 ( y != null && y.length() != 0 )
 		){
-			LonLat lonLat = new LonLat(Long.parseLong(x),Long.parseLong(y));
+			LonLat lonLat = new LonLat(Double.parseDouble(x),Double.parseDouble(y));
 			// does inverse geolocalization
 			address = addressServiceFacade.getInverseGeolocalization(request, lonLat, srid); 
 		}
